Change from Camera.Lookat to Control.target
So, when we switched to using the controls, it turns out that lookAt
stops being used to control the camera (you can, but calling
Control.update will simply ignore the current look at state). So
instead, it looks like the correct thing to do is to set the
Control.target vector. This seems to do the same thing as lookAt.
Huzzah!